85. Managing Screen Content with KeyboardAvoidingView
Concerning the behavior, here are the options:

position: this behavior is equivalent to create two Views component with an
absolute placement of the size of the keyboard. The view is juste placed
higher on the screen.

height: returns a view with total height of the screen minus the keyboard height

padding: returns a view with a paddingBottom added of the size of the
keyboard

But sadly, we have to test to really see the behavior beacuse it can
change depending on the device. trial and error